+++ src/FlaUILibrary/flaui/exception/flauierror.py
@@ -33,6 +33,7 @@ class FlaUiError(AttributeError):
     ArgumentShouldBeList = "The given argument should be an array"
     ArgumentShouldNotBeList = "The given argument should not be an array"
     PropertyNotSupported = "Property from element is not supported"
+    PropertyNotEqual = "Property value '{}' not equal to expected value '{}'"
 
     @staticmethod
     def raise_fla_ui_error(message):